Migrate fjs/cas to .f.mjs
Splits Cas, FileCas, and FileCasOperation into a sibling types.ts.
proof.f.ts migrates alongside it since its only runtime dependency is
the module itself.

The generic write() facade needed extracting into a standalone named
function (writeImpl) with a curried @template/@param/@returns JSDoc
declaration, since a bare @type on write as an object-literal property
does not introduce O1 as a body-visible type parameter for the nested
loop/publish/fail helpers — the same generic-scoping gotcha as
previous migrations, applied to a method rather than a top-level
const. A few .step() callbacks also needed explicit parameter/return
JSDoc annotations where their generic operation-set inference widened
past the declared FileCasOperation/ReadBytes.

Updates all dependents (fjs/cas/evo, fjs/cas/cli, fjs/mcp,
fjs/mcp/cas, fjs/mcp/proof, fjs/mcp/evo/proof) and stale doc references
and line-number citations across fjs/cas/README.md, fjs/cas/todo,
fjs/cas/plan, fjs/effects/todo, fjs/effects/node/todo, fjs/mcp/README.md,
fjs/todo, and todo/plan/roadmap.md.
