+++ fjs/cas/evo/module.f.ts
@@ -47,7 +47,8 @@ import type { Effect, Operation } from '../../effects/types.ts'
 import { eff } from '../../effects/eff/module.f.mjs'
 import { create, read, write } from '../../effects/memory/module.f.mjs'
 import type { Key, MemOp } from '../../effects/memory/types.ts'
-import { collectRead, type Cas } from '../module.f.ts'
+import { collectRead } from '../module.f.mjs'
+import type { Cas } from '../types.ts'
 import { cBase32ToVec, vecToCBase32 } from '../../basen/cbase32/module.f.mjs'
 import { fromVec } from '../../text/utf8/module.f.mjs'
 import { tryUtf8 } from '../../text/module.f.mjs'
