+++ fjs/cas/evo/proof.f.ts
@@ -1,6 +1,7 @@
 import { assert, assertEq } from '../../asserts/module.f.mjs'
 import { pure } from '../../effects/module.f.mjs'
-import { fileCas, type Cas } from '../module.f.ts'
+import { fileCas } from '../module.f.mjs'
+import type { Cas } from '../types.ts'
 import { sha256 } from '../../crypto/sha2/module.f.mjs'
 import { emptyState, virtual } from '../../effects/node/virtual/module.f.mjs'
 import type { Vec } from '../../types/bit_vec/types.ts'
