+++ fjs/cas/plan/staging-lease.md
@@ -519,7 +519,7 @@ a misjudged-slow uploader restarts or hands off, never corrupts.
 
 ## Relationship to the existing `casUpload` path
 
-The same migration note from `staging.md` applies: `casUpload` (`fjs/cas/module.f.ts`)
+The same migration note from `staging.md` applies: `casUpload` (`fjs/cas/module.f.mjs`)
 currently stages into `.cas/.stage/` and would be unaffected by a lease GC scanning
 `_stage/`. Migrating `casUpload` to write `<deadline>-<random256>` names under
 `_stage/` brings it under one GC scope. Unlike the lock design, the lease has no mtime
