+++ fjs/cas/plan/strategy-3.md
@@ -1,7 +1,7 @@
 # Strategy 3: Merkle Tree
 
 > **API note.** This document predates the current `Cas<O>` API in
-> `fjs/cas/module.f.ts`, which is streaming-only (`read`/`write` over
+> `fjs/cas/module.f.mjs`, which is streaming-only (`read`/`write` over
 > `List<O, IoResult<Vec>>`) and has no `fileKvStore` or whole-blob
 > `write(Vec) => hash` / `read(hash) => Vec` primitive. Below, `write`,
 > `read`, `readFile`, `writeFile`, and `fileKvStore` describe the *shape* of
