+++ fjs/cas/todo/command-architecture.md
@@ -8,7 +8,7 @@
 CAS functionality is exposed through two transports — the CLI
 (`fjs/cas/cli/module.f.ts`) and the MCP server (`fjs/mcp/module.f.ts`) —
 and each wires its own logic directly onto the store primitives in
-`fjs/cas/module.f.ts`. There is no defined notion of a *CAS command*: no
+`fjs/cas/module.f.mjs`. There is no defined notion of a *CAS command*: no
 shared command set, no typed input/output contracts, no common error
 taxonomy, and no stated policy of which transport exposes which operation.
