+++ fjs/effects/todo/effect-list-fold.md
@@ -91,8 +91,8 @@ Six, all currently strict:
 | `fjs/cas/cli/module.f.ts:66` | `c.list()` | streams once `cas.list()` does |
 | `fjs/cas/evo/module.f.ts:174` | `cas.list()` | streams once `cas.list()` does |
 | `fjs/cas/evo/module.f.ts:232` | `pure(parents)` | `fromList(parents)` |
-| `fjs/cas/module.f.ts:150` | `pure(expired)` | `fromList(expired)` |
-| `fjs/cas/module.f.ts:296` | `pure([0…7])` | `fromList([0…7])` |
+| `fjs/cas/module.f.mjs:149` | `pure(expired)` | `fromList(expired)` |
+| `fjs/cas/module.f.mjs:310` | `pure([0…7])` | `fromList([0…7])` |
 | `fjs/djs/transpiler/module.f.ts:67` | `pure(pathsArray)` | `fromList(pathsArray)` |
 
 The two `cas.list()` sites are the ones this exists for; converting them is a
