+++ fjs/effects/todo/map-step-combinator.md
@@ -32,7 +32,7 @@ Two shapes, both the same thing:
 ```
 
 Also `fjs/dev/module.f.mjs`, `fjs/cas/evo/module.f.ts`,
-`fjs/mcp/evo/module.f.ts`, `fjs/cas/module.f.ts`, `fjs/mcp/cas/module.f.ts`,
+`fjs/mcp/evo/module.f.ts`, `fjs/cas/module.f.mjs`, `fjs/mcp/cas/module.f.ts`,
 `fjs/protocol/mcp/module.f.ts`, `fjs/emergent_testing/module.f.ts`.
 
 *Constant projection* (`() => pure(v)`), overwhelmingly the "do the work, then
@@ -51,7 +51,7 @@ const program = step(
 ```
 
 Also `fjs/djs/module.f.ts`, `fjs/module.f.ts`, `fjs/ci/module.f.mjs`,
-`fjs/cas/evo/module.f.ts`, `fjs/cas/module.f.ts`, `fjs/cas/cli/module.f.ts`,
+`fjs/cas/evo/module.f.ts`, `fjs/cas/module.f.mjs`, `fjs/cas/cli/module.f.ts`,
 `fjs/mcp/cas/module.f.ts`, `fjs/protocol/mcp/module.f.ts`, `fjs/protocol/mcp/stdio/module.f.ts`,
 `fjs/emergent_testing/module.f.ts`.
