+++ fjs/mcp/evo/proof.f.ts
@@ -1,5 +1,5 @@
 import { assert, assertEq } from '../../asserts/module.f.mjs'
-import { fileCas } from '../../cas/module.f.ts'
+import { fileCas } from '../../cas/module.f.mjs'
 import { sha256 } from '../../crypto/sha2/module.f.mjs'
 import { emptyState, virtual } from '../../effects/node/virtual/module.f.mjs'
 import { vec8 } from '../../types/bit_vec/module.f.mjs'
