+++ fjs/mcp/proof.f.ts
@@ -11,7 +11,8 @@ import { msb, u8ListToVec, vec8, repeat, length, maxLengthBytes } from '../types
 import { vecToCBase32 } from '../basen/cbase32/module.f.mjs'
 import { encode as base64Encode } from '../basen/base64/module.f.mjs'
 import { utf8 } from '../text/module.f.mjs'
-import { fileCas, type FileCasOperation } from '../cas/module.f.ts'
+import { fileCas } from '../cas/module.f.mjs'
+import type { FileCasOperation } from '../cas/types.ts'
 import { dialect as revisionDialect, mediaType as revisionMediaType } from '../media/revision/module.f.ts'
 import { sha256 } from '../crypto/sha2/module.f.mjs'
 import { nonEmpty, empty as elEmpty } from '../effects/list/module.f.mjs'
